Add properly working tests for error
