+++ gourmand/coverage.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+OPBEAT_DISABLE_SEND=true coverage run --source='.' manage.py test
+coverage report -m --omit=*migrations*,*__init__.py,*tests*
